Add a separate module for htmlToDITA function
