+++ data/Britain_Complex/unit_daf.html
@@ -10,22 +10,29 @@
     </div>
     <p></p>
     <p></p>
-    <p>Excepteur officia do ipsum consequat labore minim reprehenderit in Lorem esse dolor minim exercitation voluptate. Magna ad Lorem pariatur irure laborum ad culpa sunt aliqua deserunt deserunt ut aliqua. Ea reprehenderit ullamco nostrud culpa eiusmod mollit minim dolor proident commodo minim. Cupidatat id occaecat proident irure elit Lorem non aliqua eu. Qui Lorem cillum esse do magna sit magna aliqua anim tempor exercitation voluptate. Sint cupidatat adipisicing laborum in.</p>
-    <p><ul>
+    <p>Excepteur officia do ipsum consequat labore minim reprehenderit in Lorem esse dolor minim exercitation voluptate.
+      Magna ad Lorem pariatur irure laborum ad culpa sunt aliqua deserunt deserunt ut aliqua. Ea reprehenderit ullamco
+      nostrud culpa eiusmod mollit minim dolor proident commodo minim. Cupidatat id occaecat proident irure elit Lorem
+      non aliqua eu. Qui Lorem cillum esse do magna sit magna aliqua anim tempor exercitation voluptate. Sint cupidatat
+      adipisicing laborum in.</p>
+    <p>
+    <ul>
       <li>Mode A</li>
       <li>Mode B</li>
       <li>Mode C</li>
       <li>Mode D</li>
-    </ul></p>
+    </ul>
+    </p>
     <p>
       Pellentesque habitant morbi tristique senectus et netus et malesuada fames
       ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget,
       tempor sit amet, ante. Donec eu libero sit amet quam egestas semper.
-      Aenean ultricies mi vitae est. Mauris placerat eleifend leo. This paragraph contains a <a href="./BR_Power_Units.html">link</a>
-    </p>
+      Aenean ultricies mi vitae est. Mauris placerat eleifend leo. This paragraph contains a <a
+        href="./BR_Power_Units.html">link</a>
+      <!-- </p>
     <p>
-      <h3>And now a table</h3>
-      <table border="1">
+    <h3>And now a table</h3> -->
+      <!-- <table border="1">
         <thead>
           <th>Heading one</th>
           <th>Heading two</th>
@@ -39,12 +46,12 @@
             <td colspan="2">cell three, with a <a href="./BR_Power_Units.html">link</a></td>
           </tr>
         </tbody>
-      </table>
+      </table> -->
     </p>
-    <p><img src="./Content/Images/daf_3.jpeg"/></p>
+    <p><img src="./Content/Images/daf_3.jpeg" /></p>
     <p></p>
     <div>
-      <table border="1">
+      <!-- <table border="1">
         <tr>
           <td colspan="4"><strong>Unit_DAF</strong></td>
         </tr>
@@ -76,7 +83,7 @@
           <td><strong>Max SRPM</strong></td>
           <td>0-1600</td>
         </tr>
-      </table>
+      </table> -->
     </div>
   </div>
 </div>
@@ -172,19 +179,13 @@
       are compressed to be displayed here.  But, it could be that only a subset of
       the available images are published on this page -->
       <div name="image3">
-        <a href="daf_pics.html"
-          ><img src="./Content/Images/daf_1.jpeg" width="200" height="150"
-        /></a>
+        <a href="daf_pics.html"><img src="./Content/Images/daf_1.jpeg" width="200" height="150" /></a>
       </div>
       <div name="image2">
-        <a href="daf_pics.html#number2"
-          ><img src="./Content/Images/daf_2.jpeg" width="200" height="150"
-        /></a>
+        <a href="daf_pics.html#number2"><img src="./Content/Images/daf_2.jpeg" width="200" height="150" /></a>
       </div>
       <div name="image7">
-        <a href="daf_pics.html#number3"
-          ><img src="./Content/Images/daf_3.jpeg" width="200" height="150"
-        /></a>
+        <a href="daf_pics.html#number3"><img src="./Content/Images/daf_3.jpeg" width="200" height="150" /></a>
       </div>
       <p>&nbsp;</p>
       <p>&nbsp;</p>
@@ -211,7 +212,8 @@
       Pellentesque habitant morbi tristique senectus et netus et malesuada fames
       ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget,
       tempor sit amet, ante. Donec eu libero sit amet quam egestas semper.
-      Aenean ultricies mi vitae est. Mauris placerat eleifend leo. This paragraph contains a <a href="./BR_Power_Units.html">link</a>
+      Aenean ultricies mi vitae est. Mauris placerat eleifend leo. This paragraph contains a <a
+        href="./BR_Power_Units.html">link</a>
     </p>
     <!-- <p>
       <h3>And now a table</h3>
@@ -247,7 +249,13 @@
     </ol> -->
     <p>
       <!-- <h3 id="bba_background">Background to BBA</h3> -->
-      Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum.
+      Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud
+      dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco
+      mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet
+      magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua
+      laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor ullamco mollit aliquip qui velit elit.
+      Nostrud dolor minim ut laborum aliqua laborum. Dolore mollit consequat sint amet magna dolore voluptate. Dolor
+      ullamco mollit aliquip qui velit elit. Nostrud dolor minim ut laborum aliqua laborum.
     </p>
   </div>
 </div>
