+++ data/Britain_Standalones/britain_1a.html
@@ -22,12 +22,18 @@
     <h1>PROPULSION</h1>
     <p></p>
     <p></p>
-    <p>Welcome to the propulsion section. Labore sunt Lorem sit officia incididunt do sunt aute cillum Lorem mollit amet anim occaecat. Labore nisi anim ut consectetur ut qui ea non reprehenderit amet eiusmod sit et consectetur. Lorem irure elit ex Lorem. Amet aliquip veniam voluptate anim labore reprehenderit. Ipsum officia deserunt quis deserunt duis proident ex occaecat excepteur minim mollit anim laboris.</p>
-    <p>Magna labore eu sint consectetur duis Lorem excepteur excepteur cupidatat nisi cillum ex sit. Aute culpa irure et nulla ea. Cillum esse in laborum dolor veniam exercitation mollit aliquip cillum deserunt dolor. Commodo non quis excepteur aute labore voluptate nostrud elit. Duis dolor pariatur velit consectetur consectetur aliqua ullamco et aliquip ea occaecat. Esse voluptate anim incididunt irure id cupidatat laborum do velit excepteur pariatur.</p>
-    <p></p>
-    <p></p>
+    <p>Welcome to the propulsion section. Labore sunt Lorem sit officia incididunt do sunt aute cillum Lorem mollit amet
+      anim occaecat. Labore nisi anim ut consectetur ut qui ea non reprehenderit amet eiusmod sit et consectetur. Lorem
+      irure elit ex Lorem. Amet aliquip veniam voluptate anim labore reprehenderit. Ipsum officia deserunt quis deserunt
+      duis proident ex occaecat excepteur minim mollit anim laboris.</p>
+    <p>Magna labore eu sint consectetur duis Lorem excepteur excepteur cupidatat nisi cillum ex sit. Aute culpa irure et
+      nulla ea. Cillum esse in laborum dolor veniam exercitation mollit aliquip cillum deserunt dolor. Commodo non quis
+      excepteur aute labore voluptate nostrud elit. Duis dolor pariatur velit consectetur consectetur aliqua ullamco et
+      aliquip ea occaecat. Esse voluptate anim incididunt irure id cupidatat laborum do velit excepteur pariatur.</p>
+    <!-- <p></p>
+    <p></p> -->
     <div>
-      <table border="1">
+      <!-- <table border="1">
         <tr>
           <td colspan="4"><strong>Britain 1a</strong></td>
         </tr>
@@ -59,12 +65,11 @@
           <td><strong>Max SRPM</strong></td>
           <td>1600</td>
         </tr>
-      </table>
+      </table> -->
     </div>
     <div>
-      <img src="./Content/Images/oscill_c.jpg" /><img
-        src="./Content/Images/oscill_d.jpg"
-      /><img src="./Content/Images/oscill_e.jpg" />&nbsp;&nbsp;&nbsp; Figure 1:
+      <img src="./Content/Images/oscill_c.jpg" /><img src="./Content/Images/oscill_d.jpg" /><img
+        src="./Content/Images/oscill_e.jpg" />&nbsp;&nbsp;&nbsp; Figure 1:
       oscilloscope view&nbsp;&nbsp;&nbsp; Figure 2: other oscilloscope
       view&nbsp;&nbsp;&nbsp; Figure 3: extra oscilloscope view
     </div>
@@ -151,11 +156,8 @@
           <td colspan="2">H2, 4, 21, 145</td>
           <td colspan="2">
             The quick brown fox does something similar
-            <a href="../France_Legacy/France_Legacy_Pics.html#number4"
-              >Tool screenshot a</a
-            ><a href="../France_Legacy/France_Legacy_Pics.html#number3"
-              >Tool screenshot b</a
-            >
+            <a href="../France_Legacy/France_Legacy_Pics.html#number4">Tool screenshot a</a><a
+              href="../France_Legacy/France_Legacy_Pics.html#number3">Tool screenshot b</a>
           </td>
         </tr>
         <tr>
@@ -181,13 +183,9 @@
           <td colspan="5">
             <p>
               <strong>Mode Charlie.</strong> These platforms are able to process
-              up to 3000 tonnes of processed material per day. <br /><strong
-                >Mode Delta.</strong
-              >
+              up to 3000 tonnes of processed material per day. <br /><strong>Mode Delta.</strong>
               This lower mode is most useful in reducting tentative load during
-              night time energy reduction periods. <br /><strong
-                >Note 1.</strong
-              >
+              night time energy reduction periods. <br /><strong>Note 1.</strong>
               Since the 1990s Carter Burman has has trouble generating the
               volume of spares available when installations first designed.
               Expect plant to be similar to 2016 steam powered version.
@@ -210,9 +208,7 @@
   <div id="PageLayer2">
     <h1>REMARKS</h1>
     <div name="image9" id="image9">
-      <img src="./Content/Images/oscill_c.jpg" />&nbsp;<img
-        src="./Content/Images/oscill_d.jpg"
-      />
+      <img src="./Content/Images/oscill_c.jpg" />&nbsp;<img src="./Content/Images/oscill_d.jpg" />
     </div>
     <p></p>
     <!-- <ul>
@@ -221,6 +217,8 @@
       <li>Phasellus ultrices nulla quis nibh. Quisque a lectus. Donec consectetuer ligula vulputate sem tristique cursus. Nam nulla quam, gravida non, commodo a, sodales sit amet, nisi.</li>
       <li>Pellentesque fermentum dolor. Aliquam quam lectus, facilisis auctor, ultrices ut, elementum vulputate, nunc.</li>
    </ul> -->
-   <p>Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam egestas semper. Aenean ultricies mi vitae est. Mauris placerat eleifend leo.</p>
+    <p>Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Vestibulum tortor
+      quam, feugiat vitae, ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam egestas semper. Aenean
+      ultricies mi vitae est. Mauris placerat eleifend leo.</p>
   </div>
 </div>
