+++ data/Iceland_Composite/unit_g.html
@@ -4,18 +4,24 @@
     <h1>PROPULSION</h1>
     <p></p>
     <p></p>
-    <p>Excepteur officia do ipsum consequat labore minim reprehenderit in Lorem esse dolor minim exercitation voluptate. Magna ad Lorem pariatur irure laborum ad culpa sunt aliqua deserunt deserunt ut aliqua. Ea reprehenderit ullamco nostrud culpa eiusmod mollit minim dolor proident commodo minim. Cupidatat id occaecat proident irure elit Lorem non aliqua eu. Qui Lorem cillum esse do magna sit magna aliqua anim tempor exercitation voluptate. Sint cupidatat adipisicing laborum in.</p>
-    <p><ul>
+    <p>Excepteur officia do ipsum consequat labore minim reprehenderit in Lorem esse dolor minim exercitation voluptate.
+      Magna ad Lorem pariatur irure laborum ad culpa sunt aliqua deserunt deserunt ut aliqua. Ea reprehenderit ullamco
+      nostrud culpa eiusmod mollit minim dolor proident commodo minim. Cupidatat id occaecat proident irure elit Lorem
+      non aliqua eu. Qui Lorem cillum esse do magna sit magna aliqua anim tempor exercitation voluptate. Sint cupidatat
+      adipisicing laborum in.</p>
+    <p>
+    <ul>
       <li>Mode A</li>
       <li>Mode B</li>
       <li>Mode C</li>
       <li>Mode D</li>
-    </ul></p>
+    </ul>
+    </p>
     <p></p>
     <p></p>
     <div>
       <!-- sample of prop table form unit a1 -->
-      <table border="1">
+      <!-- <table border="1">
         <tr>
           <td colspan="4"><strong>Unit_G</strong></td>
         </tr>
@@ -47,7 +53,7 @@
           <td colspan="2"><strong>Typical Transit Speed SRPM</strong></td>
           <td colspan="2">12-15</td>
         </tr>
-      </table>
+      </table> -->
     </div>
   </div>
 </div>
@@ -148,11 +154,8 @@
           <td colspan="2">H2, 4, 21, 145</td>
           <td colspan="2">
             The quick brown fox does something similar
-            <a href="../France_Legacy/France_Legacy_Pics.html#number4"
-              >Tool screenshot a</a
-            ><a href="../France_Legacy/France_Legacy_Pics.html#number3"
-              >Tool screenshot b</a
-            >
+            <a href="../France_Legacy/France_Legacy_Pics.html#number4">Tool screenshot a</a><a
+              href="../France_Legacy/France_Legacy_Pics.html#number3">Tool screenshot b</a>
           </td>
         </tr>
         <tr>
@@ -171,13 +174,13 @@
           <td colspan="2">CBR 54334</td>
         </tr>
         <tr>
-          <td><a href="Iceland_Composites_Details.html#number4">DAF</a</td>
-            <td>45.5</td>
+          <td><a href="Iceland_Composites_Details.html#number4">DAF</a< /td>
+          <td>45.5</td>
           <td colspan="2">H5-17</td>
           <td colspan="2">Follow link for juicy details</td>
         </tr>
         <tr>
-          <td><a href="Iceland_Composites_Details.html#number6">FOG</a</td>
+          <td><a href="Iceland_Composites_Details.html#number6">FOG</a< /td>
           <td>544.4</td>
           <td colspan="2">H2-12</td>
           <td colspan="2">Follow link for more juicy details</td>
@@ -207,9 +210,7 @@
   <div id="PageLayer2">
     <h1>REMARKS</h1>
     <div name="image9" id="image9">
-      <img src="./Content/Images/unit_a1.jpg" />&nbsp;<img
-        src="./Content/Images/unit_a2.jpeg"
-      />
+      <img src="./Content/Images/unit_a1.jpg" />&nbsp;<img src="./Content/Images/unit_a2.jpeg" />
     </div>
     <p></p>
     <p></p>
